+++ src/domain/usecases/predictors/GetPredictorGroupsUseCase.ts
@@ -1,6 +1,6 @@
-import { UseCase } from "../../compositionRoot";
-import { NamedRef } from "../entities/DHIS2";
-import { PredictorRepository } from "../repositories/PredictorRepository";
+import { UseCase } from "../../../compositionRoot";
+import { NamedRef } from "../../entities/DHIS2";
+import { PredictorRepository } from "../../repositories/PredictorRepository";
 
 export class GetPredictorGroupsUseCase implements UseCase {
     constructor(private predictorRepository: PredictorRepository) {}
