+++ src/domain/usecases/predictors/GetPredictorsUseCase.ts
@@ -1,7 +1,7 @@
 import { TablePagination, TableSorting } from "d2-ui-components";
-import { UseCase } from "../../compositionRoot";
-import { Predictor } from "../entities/Predictor";
-import { PredictorRepository } from "../repositories/PredictorRepository";
+import { UseCase } from "../../../compositionRoot";
+import { Predictor } from "../../entities/Predictor";
+import { PredictorRepository } from "../../repositories/PredictorRepository";
 
 interface GetPredictorsFilters {
     search?: string;
